Sprint State Changes
I Tried implementing a way to save the state of a player sprint in the context of whether the player is sprinting or not. The current setup relies on updating the walk speed on every frame while shift is held down. I only want to update the walk speed if shift is pressed or released. I'm trying to understand the logic on storing state changes so that I have an efficient and cheap way of implementing the stamina system. Right now the BP needs to have the =! check disconnected for the sprint system to work.