Gravity Gun child BP added
Added a child weapon called gravity gun, when the ray trace hits BP_Cube it will disable gravity on it and set a upward impulse making the cubes float. Each cube has it's own timer, it's pretty fun trying to juggle them and keep them in the air. I might try and add a multiplier for keeping them in the air with multiple shots if I find time.