Added Score Interface
Added a score interface that gets called off of the weapon trace. It's working and I think I can expand it fairly easy compared to doing cast and event listeners.  I still have to call all actors of class to reference the target manager within the target and am not sure if that's the best way to do it, but I'm running short on time so hopefully it works.